requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-countries==6.1.2
django-extensions==3.0.5
django-filter==2.3.0
django-webpack-loader==0.6.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pytz==2020.1
stripe==2.48.0
sqlparse==0.3.1
toml==0.10.1
zipp==3.1.0
